Add support for onnx RandomNormalLike and RandomUniformLike
Summary:
Uses a templated pattern to replace:

* RandomNormalLike with RandomNormal op.
* RandomUniformLike with RandomUniform op.

This diff also changes `op_tester` to also check that the data types of the tensors are the same.

Test Plan: Updated failing tests in workspace with additional type checks.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, matthewha

Subscribers: matthewha, evan

Maniphest Tasks: T6459

Differential Revision: https://phabricator.sourcevertex.net/D30860

+++ tests/popart/operators_test/dynamiczero_test.py
@@ -62,7 +62,7 @@ def test_dynamiczero_training(op_tester):
         mask1data[:, 8:12, 0:1, :] = 0.0
         mask0 = torch.tensor(mask0data)
         mask1 = torch.tensor(mask1data)
-        out = (((tensor * mask0) + inject_tensor) * mask1)
+        out = (((tensor * mask0) + inject_tensor) * mask1).type(torch.float32)
         outputs.append(out)
         result.append(out)
